+++ NavCog/Model/TopoMap/TopoMap.h
@@ -44,14 +44,14 @@
 
 @interface TopoMap : NSObject
 
-UIKIT_EXTERN float TopoMapUnit; // base unit is feet (1 = 1 foot = 0.3048 meter)
+UIKIT_EXTERN double TopoMapUnit; // base unit is feet (1 = 1 foot = 0.3048 meter)
 
 @property NSString* language;
 
-+ (float) unit2feet:(float)value;
-+ (float) feet2unit:(float)value;
-+ (float) unit2meter:(float)value;
-+ (float) meter2unit:(float)value;
++ (double) unit2feet:(double)value;
++ (double) feet2unit:(double)value;
++ (double) unit2meter:(double)value;
++ (double) meter2unit:(double)value;
 
 - (NavNode *)getNodeFromLayer:(NSString *)layerID withNodeID:(NSString *)nodeID;
 - (NavEdge *)getEdgeFromLayer:(NSString *)layerID withEdgeID:(NSString *)edgeID;
